Move library to ORBIT.core. Move project defaults to new defaults submodule.
