+++ ORBIT/core/library.py
@@ -41,7 +41,7 @@ from yaml import Dumper
 
 from ORBIT.core.exceptions import LibraryItemNotFoundError
 
-ROOT = os.path.abspath(os.path.join(os.path.abspath(__file__), "../.."))
+ROOT = os.path.abspath(os.path.join(os.path.abspath(__file__), "../../.."))
 default_library = os.path.join(ROOT, "library")
 
 # Need a custom loader to read in scientific notation correctly
