+++ ORBIT/phases/design/project_development.py
@@ -6,7 +6,7 @@ __maintainer__ = "Jake Nunemaker"
 __email__ = "jake.nunemaker@nrel.gov"
 
 
-from ORBIT.library import extract_library_specs
+from ORBIT.core.defaults import project_costs
 
 from .design_phase import DesignPhase
 
@@ -45,8 +45,6 @@ class ProjectDevelopment(DesignPhase):
         config = self.initialize_library(config, **kwargs)
         self.config = self.validate_config(config)
         self._outputs = {}
-
-        self.defaults = extract_library_specs("defaults", "project")
 
     def run(self):
         """
@@ -119,8 +117,8 @@ class ProjectDevelopment(DesignPhase):
         t = "site_auction_duration"
         c = "site_auction_cost"
         self._outputs["site_auction"] = {
-            "duration": development_specs.get(t, self.defaults[t]),
-            "cost": development_specs.get(c, self.defaults[c]),
+            "duration": development_specs.get(t, project_costs[t]),
+            "cost": development_specs.get(c, project_costs[c]),
         }
 
     def site_assessment_plan_development(self, **development_specs):
@@ -141,8 +139,8 @@ class ProjectDevelopment(DesignPhase):
         t = "site_assessment_plan_duration"
         c = "site_assessment_plan_cost"
         self._outputs["site_assessment_plan"] = {
-            "duration": development_specs.get(t, self.defaults[t]),
-            "cost": development_specs.get(c, self.defaults[c]),
+            "duration": development_specs.get(t, project_costs[t]),
+            "cost": development_specs.get(c, project_costs[c]),
         }
 
     def site_assessment(self, **development_specs):
@@ -162,8 +160,8 @@ class ProjectDevelopment(DesignPhase):
         t = "site_assessment_duration"
         c = "site_assessment_cost"
         self._outputs["site_assessment"] = {
-            "duration": development_specs.get(t, self.defaults[t]),
-            "cost": development_specs.get(c, self.defaults[c]),
+            "duration": development_specs.get(t, project_costs[t]),
+            "cost": development_specs.get(c, project_costs[c]),
         }
 
     def construction_operations_plan_development(self, **development_specs):
@@ -185,8 +183,8 @@ class ProjectDevelopment(DesignPhase):
         t = "construction_operations_plan_duration"
         c = "construction_operations_plan_cost"
         self._outputs["construction_operations_plan"] = {
-            "duration": development_specs.get(t, self.defaults[t]),
-            "cost": development_specs.get(c, self.defaults[c]),
+            "duration": development_specs.get(t, project_costs[t]),
+            "cost": development_specs.get(c, project_costs[c]),
         }
 
     def boem_review(self, **development_specs):
@@ -207,8 +205,8 @@ class ProjectDevelopment(DesignPhase):
         t = "boem_review_duration"
         c = "boem_review_cost"
         self._outputs["boem_review"] = {
-            "duration": development_specs.get(t, self.defaults[t]),
-            "cost": development_specs.get(c, self.defaults[c]),
+            "duration": development_specs.get(t, project_costs[t]),
+            "cost": development_specs.get(c, project_costs[c]),
         }
 
     def design_install_plan_development(self, **development_specs):
@@ -230,8 +228,8 @@ class ProjectDevelopment(DesignPhase):
         t = "design_install_plan_duration"
         c = "design_install_plan_cost"
         self._outputs["design_install_plan"] = {
-            "duration": development_specs.get(t, self.defaults[t]),
-            "cost": development_specs.get(c, self.defaults[c]),
+            "duration": development_specs.get(t, project_costs[t]),
+            "cost": development_specs.get(c, project_costs[c]),
         }
 
     @property
