+++ ORBIT/phases/design/semi_submersible_design.py
@@ -40,7 +40,6 @@ class SemiSubmersibleDesign(DesignPhase):
 
         config = self.initialize_library(config, **kwargs)
         self.config = self.validate_config(config)
-        self.extract_defaults()
         self._design = self.config.get("semisubmersible_design", {})
 
         self._outputs = {}
