+++ ORBIT/phases/design/spar_design.py
@@ -42,7 +42,6 @@ class SparDesign(DesignPhase):
 
         config = self.initialize_library(config, **kwargs)
         self.config = self.validate_config(config)
-        self.extract_defaults()
         self._design = self.config.get("spar_design", {})
 
         self._outputs = {}
