+++ ORBIT/phases/install/monopile_install/standard.py
@@ -77,7 +77,6 @@ class MonopileInstallation(InstallPhase):
 
         config = self.initialize_library(config, **kwargs)
         self.config = self.validate_config(config)
-        self.extract_defaults()
 
         self.initialize_port()
         self.initialize_wtiv()
