+++ ORBIT/phases/install/mooring_install/mooring.py
@@ -49,7 +49,6 @@ class MooringSystemInstallation(InstallPhase):
 
         config = self.initialize_library(config, **kwargs)
         self.config = self.validate_config(config)
-        self.extract_defaults()
 
         self.setup_simulation(**kwargs)
