+++ ORBIT/phases/install/oss_install/standard.py
@@ -66,7 +66,6 @@ class OffshoreSubstationInstallation(InstallPhase):
 
         config = self.initialize_library(config, **kwargs)
         self.config = self.validate_config(config)
-        self.extract_defaults()
 
         self.initialize_port()
         self.setup_simulation(**kwargs)
