+++ ORBIT/phases/install/quayside_assembly_tow/moored.py
@@ -65,7 +65,6 @@ class MooredSubInstallation(InstallPhase):
 
         config = self.initialize_library(config, **kwargs)
         self.config = self.validate_config(config)
-        self.extract_defaults()
 
         self.setup_simulation(**kwargs)
