+++ ORBIT/phases/install/scour_protection_install/standard.py
@@ -55,7 +55,6 @@ class ScourProtectionInstallation(InstallPhase):
 
         config = self.initialize_library(config, **kwargs)
         self.config = self.validate_config(config)
-        self.extract_defaults()
 
         self.setup_simulation(**kwargs)
