+++ ORBIT/phases/install/turbine_install/standard.py
@@ -84,7 +84,6 @@ class TurbineInstallation(InstallPhase):
 
         config = self.initialize_library(config, **kwargs)
         self.config = self.validate_config(config)
-        self.extract_defaults()
 
         self.initialize_port()
         self.initialize_wtiv()
