+++ tests/conftest.py
@@ -8,7 +8,7 @@ from marmot import Environment
 
 from ORBIT.core import Vessel
 from tests.data import test_weather
-from ORBIT.library import initialize_library, extract_library_specs
+from ORBIT.core.library import initialize_library, extract_library_specs
 from ORBIT.phases.install.cable_install import SimpleCable
