+++ tests/phases/design/test_project_development.py
@@ -10,7 +10,7 @@ from copy import deepcopy
 
 import pytest
 
-from ORBIT.library import extract_library_specs
+from ORBIT.core.library import extract_library_specs
 from ORBIT.phases.design import ProjectDevelopment
 
 base = {
