+++ tests/phases/install/quayside_assembly_tow/test_gravity_based.py
@@ -10,7 +10,7 @@ import pandas as pd
 import pytest
 
 from tests.data import test_weather
-from ORBIT.library import extract_library_specs
+from ORBIT.core.library import extract_library_specs
 from ORBIT.phases.install import GravityBasedInstallation
 
 config = extract_library_specs("config", "moored_install")
